requirements.txt
Adafruit-Blinka==8.68.1
Adafruit-Blinka-Raspberry-Pi5-Neopixel==1.0.0rc2
adafruit-circuitpython-busdevice==5.2.14
adafruit-circuitpython-connectionmanager==3.1.6
adafruit-circuitpython-dht==4.0.10
adafruit-circuitpython-requests==4.1.15
adafruit-circuitpython-typing==1.12.3
Adafruit-PlatformDetect==3.85.0
Adafruit-PureIO==1.1.11
aenum==3.1.16
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.4.0
binho-host-adapter==0.1.6
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorzero==2.0
Deprecated==1.3.1
Flask==3.1.2
frozenlist==1.8.0
future==1.0.0
gpiozero==2.0.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
lgpio==0.2.2.0
line-bot-sdk==3.21.0
MarkupSafe==3.0.3
multidict==6.7.0
propcache==0.4.1
pydantic==2.12.5
pydantic_core==2.41.5
pyftdi==0.57.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pyusb==1.3.1
requests==2.32.5
RPi.GPIO==0.7.1
rpi_ws281x==5.0.0
RPLCD==1.4.0
six==1.17.0
smbus2==0.6.0
sysv_ipc==1.1.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
Werkzeug==3.1.4
wrapt==2.0.1
yarl==1.22.0
